python-agent/pyproject.toml
[project]
name = "beyondpresence-livekit-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "livekit>=0.13.1",
    "livekit-agents>=0.10.3",
    "livekit-plugins-openai>=0.9.0",
    "livekit-plugins-bey>=0.1.0",
    "livekit-plugins-deepgram>=0.6.12",
    "livekit-plugins-silero>=0.7.4",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.9.5",
]

[tool.uv]
dev-dependencies = []